refactor : 도메인 중심 폴더 구조 리팩토링
