+++ app/[universityName]/[lectureId]/components/index.ts
@@ -1,3 +1,2 @@
 export { default as LectureInfo } from './LectureInfo';
 export { default as ReviewList } from './ReviewList';
-export { default as LectureInfoSkeleton } from './LectureInfoSkeleton';
