--- components/common/card/skeleton/index.ts
@@ -1,4 +1,0 @@
-export { default as CardSkeleton } from './CardSkeleton';
-export { default as CardListSkeleton } from './CardListSkeleton';
-export { default as ReviewCardSkeleton } from './ReviewCardSkeleton';
-export { default as ReviewCardListSkeleton } from './ReviewCardListSkeleton';
