+++ components/common/index.ts
@@ -1,5 +1,3 @@
-export { default as LectureCard } from './card/LectureCard';
-export { default as ReviewCard } from './card/ReviewCard';
 export { default as StarRating } from './starRating/StarRating';
 export { default as Selector } from './Selector/Selector';
 export { default as Modal } from './modal/Modal';
