+++ domains/lecture/skeleton/CardListSkeleton.tsx
