+++ domains/lecture/skeleton/CardSkeleton.tsx
