+++ domains/lecture/skeleton/LectureInfoSkeleton.tsx
@@ -1,4 +1,4 @@
-import styles from './styles/LectureInfoSkeleton.module.css';
+import styles from './LectureInfoSkeleton.module.css';
 
 export default function LectureInfoSkeleton() {
   return (
