+++ domains/lecture/styles/LectureCard.module.css
