+++ public/index.ts
@@ -1,0 +1,2 @@
+export { default as manImage } from './assets/default-profile/man.png';
+export { default as womanImage } from './assets/default-profile/woman.png';
